[AIRFLOW-5807] Move SFTP from contrib to core
