+++ airflow/sensors/sftp_sensor.py
@@ -16,10 +16,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
+"""
+This module contains SFTP sensor.
+"""
 from paramiko import SFTP_NO_SUCH_FILE
 
-from airflow.contrib.hooks.sftp_hook import SFTPHook
+from airflow.hooks.sftp_hook import SFTPHook
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
